Introduce stable versioned configs for RHTAP
The motivation is to make it so the RHTAP sample templates do not
always pull in the latest rego policies from main branch of
ec-policies.

Because RHTAP users might not always be using the latest version of
the ec cli, there's a chance that an update to the policies might
not be backwards compatible, which could break RHTAP useres. So
that's why it's better if RHTAP users are not using main branch
ec-policies.

The versions are based on the downstream RH builds of EC, which are
what gets shipped with RHTAS and hence with RHTAP.

Ref: https://issues.redhat.com/browse/EC-605
See also: https://issues.redhat.com/browse/EC-695